+++ Plutarch/DataRepr/Internal.hs
@@ -8,8 +8,10 @@ module Plutarch.DataRepr.Internal (
   punDataSum,
   ptryIndexDataSum,
   pmatchDataSum,
+  pdcons,
+  pdnil,
   DataReprHandlers (..),
-  PDataRecord,
+  PDataRecord (..),
   PLabeledType (..),
   type PLabelIndex,
   type PUnLabel,
@@ -23,7 +25,7 @@ module Plutarch.DataRepr.Internal (
 
 import Data.Kind (Type)
 import Data.List (groupBy, maximumBy, sortOn)
-import Data.Proxy (Proxy)
+import Data.Proxy (Proxy (Proxy))
 import GHC.TypeLits (
   ErrorMessage (ShowType, Text, (:<>:)),
   KnownNat,
@@ -33,21 +35,41 @@ import GHC.TypeLits (
   natVal,
   type (+),
  )
-import Generics.SOP (Code, Generic, I (I), NP (Nil, (:*)), SOP (SOP), to)
+import Generics.SOP (
+  All,
+  All2,
+  AllZipN,
+  Code,
+  Generic,
+  I (I),
+  K (K),
+  LiftedCoercible,
+  NP (Nil, (:*)),
+  POP,
+  SListI,
+  SOP (SOP),
+  from,
+  hcmap,
+  hcollapse,
+  hfromI,
+  hindex,
+  to,
+ )
 import Plutarch (
   Dig,
+  PInner,
   PMatch,
   PType,
+  PlutusType,
   S,
   Term,
-  TermCont,
-  hashOpenTerm,
+  pcon',
   perror,
   phoistAcyclic,
   plam,
   plet,
-  pmatch,
-  runTermCont,
+  pmatch',
+  pto,
   (#),
   (#$),
   type (:-->),
@@ -56,9 +78,11 @@ import Plutarch.Bool (pif, (#==))
 import Plutarch.Builtin (
   PAsData,
   PBuiltinList,
+  PBuiltinPair,
   PData,
   PIsData,
   pasConstr,
+  pconstrBuiltin,
   pdata,
   pforgetData,
   pfromData,
@@ -69,15 +93,59 @@ import Plutarch.DataRepr.Internal.Generic (MkSum (mkSum))
 import Plutarch.DataRepr.Internal.HList (type Drop, type IndexList)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal (S (SI))
-import Plutarch.Lift (PConstant, PConstantRepr, PConstanted, PLift, pconstantFromRepr, pconstantToRepr)
-import Plutarch.List (pdrop, ptryIndex)
+import Plutarch.Internal.TypeFamily (ToPType2)
+import Plutarch.Lift (PConstant, PConstantRepr, PConstanted, PLift, pconstant, pconstantFromRepr, pconstantToRepr)
+import Plutarch.List (PListLike (pnil), pcons, pdrop, phead, ptail, ptryIndex)
+import Plutarch.TermCont (TermCont, hashOpenTerm, runTermCont)
 import Plutarch.Unsafe (punsafeCoerce)
 import qualified Plutus.V1.Ledger.Api as Ledger
 
 {- | A "record" of `exists a. PAsData a`. The underlying representation is
  `PBuiltinList PData`.
 -}
-data PDataRecord (as :: [PLabeledType]) (s :: S)
+data PDataRecord (as :: [PLabeledType]) (s :: S) where
+  PDCons ::
+    forall name x xs s.
+    Term s (PAsData x) ->
+    (Term s (PDataRecord xs)) ->
+    PDataRecord ((name ':= x) ': xs) s
+  PDNil :: PDataRecord '[] s
+
+instance PlutusType (PDataRecord ((name ':= x) ': xs)) where
+  type PInner (PDataRecord ((name ':= x) ': xs)) _ = PBuiltinList PData
+  pcon' (PDCons x xs) = pto result
+    where
+      result :: Term _ (PDataRecord ((name ':= x) ': xs))
+      result = pdcons # x # xs
+  pmatch' l' f = plet l' $ \l ->
+    let x :: Term _ (PAsData x)
+        x = punsafeCoerce $ phead # l
+        xs :: Term _ (PDataRecord xs)
+        xs = punsafeCoerce $ ptail # l
+     in f $ PDCons x xs
+
+instance PlutusType (PDataRecord '[]) where
+  type PInner (PDataRecord '[]) _ = PBuiltinList PData
+  pcon' PDNil = pnil
+  pmatch' _ f = f PDNil
+
+{- | Cons a field to a data record.
+
+You can specify the label to associate with the field using type applications-
+
+@
+
+foo :: Term s (PDataRecord '[ "fooField" ':= PByteString ])
+foo = pdcons @"fooField" # pdata (phexByteStr "ab") # pdnil
+
+@
+-}
+pdcons :: forall label a l s. Term s (PAsData a :--> PDataRecord l :--> PDataRecord ((label ':= a) ': l))
+pdcons = punsafeCoerce $ pcons @PBuiltinList @PData
+
+-- | An empty 'PDataRecord'.
+pdnil :: Term s (PDataRecord '[])
+pdnil = punsafeCoerce $ pnil @PBuiltinList @PData
 
 data PLabeledType = Symbol := PType
 
@@ -205,9 +273,28 @@ pmatchDataSum d handlers =
 -}
 newtype PIsDataReprInstances (a :: PType) (s :: S) = PIsDataReprInstances (a s)
 
+-- TODO: This 'PMatch' constraint needs to be changed to 'PlutusType (breaking change).
 class (PMatch a, PIsData a) => PIsDataRepr (a :: PType) where
   type PIsDataReprRepr a :: [[PLabeledType]]
   type PIsDataReprRepr a = PDataRecordFields2 (Code (a 'SI))
+
+  pconRepr :: a s -> Term s (PDataSum (PIsDataReprRepr a))
+  default pconRepr ::
+    forall s code pcode.
+    ( Generic (a s)
+    , code ~ Code (a s)
+    , pcode ~ ToPType2 code
+    , All SListI pcode
+    , All Singleton code
+    , All2 IsBuiltinList pcode
+    , AllZipN POP (LiftedCoercible I (Term s)) code pcode
+    ) =>
+    a s ->
+    Term s (PDataSum (PIsDataReprRepr a))
+  pconRepr x = punsafeCoerce expected
+    where
+      expected :: Term _ (PAsData (PBuiltinPair PInteger (PBuiltinList PData)))
+      expected = gpconRepr @a $ from x
 
   pmatchRepr :: forall s b. Term s (PDataSum (PIsDataReprRepr a)) -> (a s -> Term s b) -> Term s b
   default pmatchRepr ::
@@ -221,6 +308,23 @@ class (PMatch a, PIsData a) => PIsDataRepr (a :: PType) where
     Term s b
   pmatchRepr dat =
     pmatchDataSum dat . mkDataReprHandler @s @a @0 @code
+
+gpconRepr ::
+  forall a s code pcode.
+  ( Generic (a s)
+  , code ~ Code (a s)
+  , pcode ~ ToPType2 code
+  , All SListI pcode
+  , All Singleton code
+  , All2 IsBuiltinList pcode
+  , AllZipN POP (LiftedCoercible I (Term s)) code pcode
+  ) =>
+  SOP I (Code (a s)) ->
+  Term s (PAsData (PBuiltinPair PInteger (PBuiltinList PData)))
+gpconRepr x = pconstrBuiltin # pconstant (toInteger $ hindex sop) # head (hcollapse sop)
+  where
+    sop :: SOP (K (Term s (PBuiltinList PData))) pcode
+    sop = hcmap (Proxy @IsBuiltinList) (K . dataListFrom) $ hfromI x
 
 -- | Create a `DataReprhandlers` starting from `n`th sum constructor
 class MkDataReprHandler (s :: S) (a :: PType) (n :: Nat) (rest :: [[Type]]) where
@@ -255,8 +359,10 @@ instance PIsDataRepr a => PIsData (PIsDataReprInstances a) where
   pdata = punsafeCoerce
   pfromData = punsafeCoerce
 
-instance PIsDataRepr a => PMatch (PIsDataReprInstances a) where
-  pmatch x f = pmatchRepr (punsafeCoerce x) (f . PIsDataReprInstances)
+instance PIsDataRepr a => PlutusType (PIsDataReprInstances a) where
+  type PInner (PIsDataReprInstances a) _ = PDataSum (PIsDataReprRepr a)
+  pcon' (PIsDataReprInstances x) = pconRepr x
+  pmatch' x f = pmatchRepr x (f . PIsDataReprInstances)
 
 newtype DerivePConstantViaData (h :: Type) (p :: PType) = DerivePConstantViaData h
 
@@ -265,3 +371,13 @@ instance (PIsDataRepr p, PLift p, Ledger.FromData h, Ledger.ToData h) => PConsta
   type PConstanted (DerivePConstantViaData h p) = p
   pconstantToRepr (DerivePConstantViaData x) = Ledger.toData x
   pconstantFromRepr x = DerivePConstantViaData <$> Ledger.fromData x
+
+-- I wish type families could be applied partially....
+class Singleton a
+instance Singleton (x ': '[])
+
+class IsBuiltinList a where
+  dataListFrom :: Term s a -> Term s (PBuiltinList PData)
+
+instance IsBuiltinList (PDataRecord l) where
+  dataListFrom = punsafeCoerce