+++ Plutarch/DataRepr/Internal/HList.hs
@@ -22,28 +22,26 @@ module Plutarch.DataRepr.Internal.HList (
 ) where
 
 import Data.Kind (Type)
-import GHC.Records (HasField (..))
-import GHC.TypeLits (
-  ErrorMessage (Text),
-  Nat,
-  Symbol,
-  TypeError,
-  type (-),
+import GHC.Records (HasField, getField)
+import GHC.TypeLits (Symbol)
+import Plutarch (Term)
+import Plutarch.Builtin
+import Plutarch.DataRepr.Internal.FromData (PFromDataable, pmaybeFromAsData)
+import Plutarch.DataRepr.Internal.HList.Utils (
+  Drop,
+  Elem (Here, There),
+  IndexLabel,
+  IndexList,
+  Labeled (Labeled, unLabeled),
+  SingleItem,
  )
 
 --------------------------------------------------------------------------------
 ---------- HList and HRec types
 
-newtype Labeled sym a = Labeled {unLabeled :: a}
-
 data HRec (as :: [Type]) where
   HNil :: HRec '[]
   HCons :: (Labeled name a) -> HRec as -> HRec ((Labeled name a) ': as)
-
--- | GADT proof-witness of HList membership, usable as an index
-data Elem (a :: k) (as :: [k]) where
-  Here :: Elem a (a ': as)
-  There :: Elem a as -> Elem a (b ': as)
 
 ---------- Field indexing functions
 
@@ -60,36 +58,14 @@ indexHRec HNil impossible = case impossible of {}
   >>> hrecField @"y" @["x", "y", "z"] xs
   >>> 2
 -}
-hrecField ::
+hrecField' ::
   forall name a as.
   ( (IndexLabel name as ~ a)
   , ElemOf name a as
   ) =>
   HRec as ->
   a
-hrecField xs = unLabeled $ indexHRec xs $ elemOf @name @a @as
-
----------- Type families
-
--- | Indexing type-level lists
-type family IndexList (n :: Nat) (l :: [k]) :: k where
-  IndexList _ '[] = TypeError ( 'Text "IndexList: index out of bounds")
-  IndexList 0 (x ': _) = x
-  IndexList n (x : xs) = IndexList (n - 1) xs
-
--- | Indexing list of labeled types by label
-type family IndexLabel (name :: Symbol) (as :: [Type]) :: Type where
-  IndexLabel name ((Labeled name a) ': _) = a
-  IndexLabel name (_ ': as) = IndexLabel name as
-
--- | Return the single item from a singleton list
-type family SingleItem (as :: [k]) :: k where
-  SingleItem '[a] = a
-
--- | Drop first n fields of a list
-type family Drop (n :: Nat) (as :: [k]) :: [k] where
-  Drop 0 xs = xs
-  Drop n (x ': xs) = Drop (n - 1) xs
+hrecField' xs = unLabeled $ indexHRec xs $ elemOf @name @a @as
 
 ---------- Internal utils
 
@@ -129,13 +105,33 @@ instance
   elemOf :: Elem (Labeled name a) (b ': as)
   elemOf = There (elemOf @name @a @as)
 
----------- HasField instances
+{- |
+  Index a `HRec` with a field in a provided list of data fields.
+  Implicitly unwraps `PAsData a` to `a` when necessary.
 
-instance
-  forall name a as.
-  ( (IndexLabel name as) ~ a
+  >>> xs = HRec @["x", "y", "z"] (HCons 1 (HCons 2 (HCons 3 HNil)))
+  >>> hrecField @"y" @["x", "y", "z"] xs
+  >>> 2
+-}
+hrecField ::
+  forall name a as b c s.
+  ( IndexLabel name as ~ a
   , ElemOf name a as
+  , Term s (PAsData b) ~ a
+  , PFromDataable b c
   ) =>
-  HasField name (HRec as) a
+  HRec as ->
+  Term s c
+hrecField xs = pmaybeFromAsData $ hrecField' @name xs
+
+---------- HasField instances
+instance
+  forall name a as b c s.
+  ( IndexLabel name as ~ a
+  , ElemOf name a as
+  , Term s (PAsData b) ~ a
+  , PFromDataable b c
+  ) =>
+  HasField name (HRec as) (Term s c)
   where
   getField = hrecField @name